ensure stat returns a data.frame
The next version of ggplot2 will move to using vctrs internally. Because of this, returning a matrix rather than a data.frame in a stat will cause an error downstream. This was never really supported but apparently worked by accident :-)

We plan to release ggplot2 by the end of October and hope a fix in ggpubr have been released at that time

thank you for your contributions to the ecosystem

kind regards
Thomas